Move rollDice to new DiceService
Fixes #1
